github-bot/requirements.txt
fastapi
uvicorn
octokitpy
python-multipart